llm-service/requirements.txt
# Core dependencies
fastapi==0.110.0
uvicorn[standard]==0.27.1
gunicorn==21.2.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1

# Database
pymysql==1.1.0
sqlalchemy==2.0.27
cryptography==42.0.5

# LlamaIndex and RAG dependencies
llama-index==0.10.12
llama-index-embeddings-huggingface==0.1.4
llama-index-llms-llama-cpp==0.1.3
llama-index-vector-stores-chroma==0.1.6
chromadb==0.4.22

# LLM dependencies
llama-cpp-python==0.2.52
transformers==4.38.1
sentence-transformers==2.5.1
torch==2.2.0

# Utilities
pydantic==2.6.1
httpx==0.26.0
aiofiles==23.2.1
pandas==2.2.0
numpy==1.26.4

# Document processing
pypdf==4.0.1
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23